+++ src/components/hooks/helper/useLoadingHelper.test.tsx
@@ -1,10 +1,10 @@
 import { renderHook } from "@testing-library/react"
 
-import { useLoading } from "@/components/hooks/useLoading"
+import { useLoadingHelper } from "@/components/hooks/helper/useLoadingHelper"
 
 describe("useLoading", () => {
   test("初期値としてfalseが設定されていること", () => {
-    const { result } = renderHook(() => useLoading())
+    const { result } = renderHook(() => useLoadingHelper())
     expect(result.current.isLoading).toBe(false)
   })
 })
