+++ src/components/hooks/helper/usePrefectureHelper.test.tsx
@@ -1,6 +1,6 @@
 import { act, renderHook } from "@testing-library/react"
 
-import { usePrefectureHelper } from "@/components/hooks/usePrefectureHelper"
+import { usePrefectureHelper } from "@/components/hooks/helper/usePrefectureHelper"
 import { prefListFixture } from "@/test/fixtures/prefListFixture"
 
 describe("usePrefectureHelper", () => {
