+++ src/components/presentational/PrefCheckboxes.tsx
@@ -1,7 +1,7 @@
 import { FC } from "react"
 import { TabList, TabPanel, Tabs } from "react-tabs"
 
-import { usePrefectureHelper } from "@/components/hooks/usePrefectureHelper"
+import { usePrefectureHelper } from "@/components/hooks/helper/usePrefectureHelper"
 import { CustomTab } from "@/components/presentational/elements/CustomTab"
 import { PrefCheckbox } from "@/components/presentational/PrefCheckbox"
 import { PrefCode, PrefCodeList, PrefList, PrefName } from "@/types/pref"
